agenda/requirements.txt
django==4.1
django-crispy-forms==1.13
django-colorfield
djangorestframework-simplejwt
pyjwt